tools/inferflow-cli/pyproject.toml
[project]
name = "inferflow-cli"
version = "0.1.0"
description = "Inferflow developer CLI"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1",
    "httpx>=0.28",
    "kubernetes>=30.1",
    "pydantic>=2.11",
    "python-dotenv>=1.1",
    "pyyaml>=6.0",
    "questionary>=2.1",
    "rich>=13.9",
    "typer>=0.12",
]

[project.scripts]
inferflow = "inferflow_cli.main:app"

[tool.uv]
package = true
